+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## 1.10.0.0
 
-* Changed the type of the Lenses ppMaxBBSizeL, ppMaxTxSizeL, ppMaxBHSizeL, ppuMaxBBSizeL, ppuMaxTxSizeL and ppuMaxBHSizeL
+* Add `getWitsVKeyNeeded` to `EraUTxO`
+* Changed the type of the lenses `ppMaxBBSizeL`, `ppMaxTxSizeL`, `ppMaxBHSizeL`,
+  `ppuMaxBBSizeL`, `ppuMaxTxSizeL` and `ppuMaxBHSizeL`
 * Fix `ToJSON`/`FromJSON` for `CostModels`. Make sure that `CostModels` can roundtrip
   through JSON. Also report `CostModels` failures in JSON.
 * Add `costModelParamsCount`